Rename package zkp-stark
