+++ crypto/zkp-stark/examples/vdf.rs
@@ -2,7 +2,7 @@
 use env_logger;
 use log::info;
 use macros_decl::field_element;
-use openstark::{Constraints, Provable, RationalExpression, TraceTable, Verifiable};
+use zkp_stark::{Constraints, Provable, RationalExpression, TraceTable, Verifiable};
 use primefield::FieldElement;
 use std::time::Instant;
 use u256::U256;
