+++ crypto/zkp-stark/src/proof_of_work.rs
