+++ crypto/zkp-stark/src/trace_table.rs
